+++ R/aaa02_broadcast_casting.R
@@ -1,14 +1,33 @@
 #' Details on Casting Functions
 #' 
 #' @description
-#' This help page gives some additional information on the casting functions
-#' provided by the 'broadcast' package. \cr
+#' 'broadcast' provides several "casting" functions. \cr
+#' These can facility complex forms of broadcasting that would normally not be possible. \cr
+#' But these "casting" functions also have their own merit, beside empowering complex broadcasting. \cr
 #' \cr
+#' The following casting functions are available:
+#' 
+#'  - \link{acast}: \cr
+#'  casts group-based subset of an array into a new dimension. \cr
+#'  Useful for, for example, computing \bold{grouped} broadcasted operations. \cr
+#'  - \link{cast_hier2dim}: \cr
+#'  casts a nested/hierarchical list into a dimensional list (i.e. recursive array). \cr
+#'  Useful because one cannot broadcast through nesting, but one \bold{can} broadcast along dimensions.
+#'  - \link{cast_dim2hier}: \cr
+#'  casts a dimensional list into a nested/hierarchical list; the opposite of \link{cast_hier2dim}. \cr
+#'  - \link{cast_dim2flat}: \cr
+#'  casts a dimensional list into a flattened list, but with names that indicate their original dimensional positions. \cr
+#'  Mostly useful for printing or summarizing dimensional lists.
+#'  - \link{dropnests}: \cr
+#'  drop redundant nesting in lists; mostly used for facilitating the above casting functions. \cr \cr
 #' 
 #' 
-#' @section Argument `in2out = TRUE`: 
+#' @section Shared Argument `in2out`: 
 #' The \link{hier2dim}, \link{cast_hier2dim}, and \link{cast_dim2hier} methods
 #' all have the `in2out` argument. \cr
+#' \cr
+#' \cr
+#' \bold{`in2out = TRUE`} \cr
 #' By default `in2out` is `TRUE`. \cr
 #' This means the call \cr
 #' `y <- cast_hier2dim(x)` \cr
@@ -29,13 +48,10 @@
 #' `y <- cast_dim2hier(x)`. \cr
 #' Then it holds that: \cr
 #' `x[[i]][[j]][[k]]` corresponds to `y[[k, j, i]]`, \cr
-#' \eqn{\forall}(`i`, `j`, `k`) , provided `x[[i]][[j]][[k]]` exists. \cr \cr
-#' 
-#' 
-#' 
-#' @section Argument `in2out = FALSE`:
-#' The \link{hier2dim}, \link{cast_hier2dim}, and \link{cast_dim2hier} methods
-#' all have the `in2out` argument. \cr
+#' \eqn{\forall}(`i`, `j`, `k`) , provided `x[[i]][[j]][[k]]` exists. \cr
+#' \cr
+#' \cr
+#' \bold{`in2out = FALSE`} \cr
 #' If `in2out = FALSE`, the call \cr
 #' `y <- cast_hier2dim(x, in2out = FALSE)` \cr
 #' will cast the surface-level elements of `x` to the rows of `y`,