+++ R/overload_ops_andor.R
@@ -18,7 +18,6 @@
   else if(is.raw(e1) && is.raw(e2)) {
     out <- .bc_bit(e1, e2, 1L, sys.call())
   }
-  
   else {
     stop("operations are possible only for numeric, logical or complex types")
   }