+++ inst/examples/bc_ifelse.R
@@ -1,5 +1,5 @@
 
-x.dim <- c(5, 3,2)
+x.dim <- c(5, 3, 2)
 x.len <- prod(x.dim)
 
 x <- array(sample(1:100), x.dim)