+++ inst/examples/bc_rel.R
@@ -1,0 +1,14 @@
+
+x.dim <- c(4:2)
+x.len <- prod(x.dim)
+x.data <- sample(c(NA, 1.1:1000.1), x.len, TRUE)
+x <- array(x.data, x.dim)
+y <- array(1:50, c(4,1,1))
+
+
+bc.rel(x, y, "==")
+bc.rel(x, y, "!=")
+bc.rel(x, y, "<")
+bc.rel(x, y, ">")
+bc.rel(x, y, "<=")
+bc.rel(x, y, ">=")