+++ inst/tinytest/overload/test-relops_equneq_only.R
@@ -43,16 +43,18 @@ accept_test <- function(x, y) {
 
 # equals ====
 nres <- 5 * 6 * 6 * 3 * 3 # number of tests performed here
-expected <- out <- vector("list", nres)
+expected <- out1 <- out2 <- vector("list", nres)
 basefun <- function(x, y) {
   out <- x == y
 
   return(out)
 }
-testfun <- function(x, y) {
+testfun1 <- function(x, y) {
   ab(x) == ab(y)
 }
-
+testfun2 <- function(x, y) {
+  bc.rel(x, y, "==")
+}
 
 i <- 1L
 for(iSample in 1:5) { # re-do tests with different random configurations
@@ -105,27 +107,27 @@ for(iSample in 1:5) { # re-do tests with different random configurations
               expected[[i]] <- basefun(drop(x), drop(y))
               
               # attributes(expected[[i]]) <- NULL # must be a vector if tdim == NULL
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             else if(length(y) == 1L && length(x) == 1L) {
               # CASE 2: x and y are both scalar arrays
               expected[[i]] <- basefun(as.vector(x), as.vector(y))
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             else if(length(x) == 1L && length(y) > 1L) {
               # CASE 3: x is scalar, y is not
               expected[[i]] <- basefun(as.vector(x), rep_dim(y, tdim))
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             else if(length(y) == 1L && length(x) > 1L) {
               # CASE 4: y is scalar, x is not
               expected[[i]] <- basefun(rep_dim(x, tdim), as.vector(y))
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             else {
               # CASE 5: x and y are both non-reducible arrays
               expected[[i]] <- basefun(rep_dim(x, tdim), rep_dim(y, tdim))
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             # END CASES
             
@@ -135,15 +137,19 @@ for(iSample in 1:5) { # re-do tests with different random configurations
             # the following code is meant to ensure NaN results turn to NA, like 'broadcast' does
             ind.NaN <- is.nan(expected[[i]])
             expected[[i]][ind.NaN] <- .return_missing(expected[[i]][ind.NaN])
-            ind.NaN <- is.nan(out[[i]])
-            out[[i]][ind.NaN] <- .return_missing(out[[i]][ind.NaN])
+            
+            ind.NaN <- is.nan(out1[[i]])
+            out1[[i]][ind.NaN] <- .return_missing(out1[[i]][ind.NaN])
+            
+            ind.NaN <- is.nan(out2[[i]])
+            out2[[i]][ind.NaN] <- .return_missing(out2[[i]][ind.NaN])
             
             dim(expected[[i]]) <- tdim
             
-            out[[i]] <- unclass(out[[i]]) # because broadcaster attribute is preserved
+            out1[[i]] <- unclass(out1[[i]]) # because broadcaster attribute is preserved
             
             
-            i <- i + 1L
+            i <- i + 2L
           }
         }
       }
@@ -153,23 +159,28 @@ for(iSample in 1:5) { # re-do tests with different random configurations
 enumerate <- enumerate + i # count number of tests
 # test results:
 expect_equal(
-  expected, out
+  expected, out1
+)
+expect_equal(
+  expected, out2
 )
 
 
 
 # unequals ====
 nres <- 5 * 6 * 6 * 3 * 3 # number of tests performed here
-expected <- out <- vector("list", nres)
+expected <- out1 <- out2 <- vector("list", nres)
 basefun <- function(x, y) {
   out <- x != y
   
   return(out)
 }
-testfun <- function(x, y) {
+testfun1 <- function(x, y) {
   ab(x) != ab(y)
 }
-
+testfun2 <- function(x, y) {
+  bc.rel(x, y, "!=")
+}
 
 i <- 1L
 for(iSample in 1:5) { # re-do tests with different random configurations
@@ -222,27 +233,27 @@ for(iSample in 1:5) { # re-do tests with different random configurations
               expected[[i]] <- basefun(drop(x), drop(y))
               
               # attributes(expected[[i]]) <- NULL # must be a vector if tdim == NULL
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             else if(length(y) == 1L && length(x) == 1L) {
               # CASE 2: x and y are both scalar arrays
               expected[[i]] <- basefun(as.vector(x), as.vector(y))
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             else if(length(x) == 1L && length(y) > 1L) {
               # CASE 3: x is scalar, y is not
               expected[[i]] <- basefun(as.vector(x), rep_dim(y, tdim))
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             else if(length(y) == 1L && length(x) > 1L) {
               # CASE 4: y is scalar, x is not
               expected[[i]] <- basefun(rep_dim(x, tdim), as.vector(y))
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             else {
               # CASE 5: x and y are both non-reducible arrays
               expected[[i]] <- basefun(rep_dim(x, tdim), rep_dim(y, tdim))
-              out[[i]] <- testfun(x, y)
+              out1[[i]] <- testfun1(x, y); out2[[i]] <- testfun2(x, y)
             }
             # END CASES
             
@@ -252,14 +263,19 @@ for(iSample in 1:5) { # re-do tests with different random configurations
             # the following code is meant to ensure NaN results turn to NA, like 'broadcast' does
             ind.NaN <- is.nan(expected[[i]])
             expected[[i]][ind.NaN] <- .return_missing(expected[[i]][ind.NaN])
-            ind.NaN <- is.nan(out[[i]])
-            out[[i]][ind.NaN] <- .return_missing(out[[i]][ind.NaN])
+            
+            ind.NaN <- is.nan(out1[[i]])
+            out1[[i]][ind.NaN] <- .return_missing(out1[[i]][ind.NaN])
+            
+            ind.NaN <- is.nan(out2[[i]])
+            out2[[i]][ind.NaN] <- .return_missing(out2[[i]][ind.NaN])
             
             dim(expected[[i]]) <- tdim
             
-            out[[i]] <- unclass(out[[i]]) # because broadcaster attribute is preserved
+            out1[[i]] <- unclass(out1[[i]]) # because broadcaster attribute is preserved
             
-            i <- i + 1L
+            
+            i <- i + 2L
           }
         }
       }
@@ -269,7 +285,8 @@ for(iSample in 1:5) { # re-do tests with different random configurations
 enumerate <- enumerate + i # count number of tests
 # test results:
 expect_equal(
-  expected, out
+  expected, out1
 )
-
-
+expect_equal(
+  expected, out2
+)